requirements.txt
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
boto3==1.39.11
botocore==1.39.11
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
fastapi==0.116.1
filelock==3.18.0
fonttools==4.59.0
fsspec==2025.7.0
h11==0.16.0
hydra-core==1.3.2
idna==3.10
iopath==0.1.10
Jinja2==3.1.6
jmespath==1.0.1
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
omegaconf==2.3.0
opencv-python==4.12.0.88
packaging==25.0
pillow==11.3.0
portalocker==3.2.0
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
s3transfer==0.13.1
SAM-2 @ git+https://github.com/facebookresearch/sam2.git@2b90b9f5ceec907a1c18123530e92e794ad901a4
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
starlette==0.47.2
sympy==1.14.0
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
